Restore original tun.go code
